add listStageFiles per domain
